+++ test/controllers/tags_controller_test.rb
@@ -1,4 +1,4 @@
-require "test_helper"
+require 'test_helper'
 
 class TagsControllerTest < ActionDispatch::IntegrationTest
   # test "the truth" do
